Completed the planet generator
I completed the planet generator. It will now create a list of planet objects based on a fraction of the galaxy's star count. It will then choose a size for the planet and a distance between 0.4 and 2.0 AU. The primary biome is determined from the distance from the star. For instance, a planet close to the star is likely to be a lava, barren, or toxic world, whereas a planet further away is likely to be a barren or arctic world.

I also created the galaxy generator that creates a galaxy object and gives it the number of stars. This is a slimmed down version from what it was originally meant to do, but I'm keeping the object framework just in case I want to revisit it later and implement the missing features/expand on this at a later date.