Started the Species Generator
I created the species class and added lists to choose from for number of limbs and general morphology. Currently, there are only three main morphologies for testing purposes but it is easily expandable.

The program also chooses the likelyhood of of the morphology based on the main planet biome. For instance, a forest world is more likely to produce a primate or arthropod species, while an ocean world will be a cetacean.