Added Pandas and Exported DataFrames as CSV
I added pandas to start using DataFrames. I'll need that to visualize the data. I think it's still loosely tied to sqlite somehow? But I'm not 100% on that. I'm really going to have to trim the fat on this at the end.

Also, I lowered the number of stars in the galaxy drastically. It was taking several minutes to generate the planet and species DataFrames as it was. I should still have a large enough sample size to get some good data correlations though