Implemented Seaborn for Visualization
Began to implement seaborn for data visualization. The next step is to maybe add a few more possibilities for morphologies. I would also like to associate morphology with height.

Trimmed a lot of the fat, and removed all referenced to sqlite from main.py as it was no longer necessary.

I'm also slightly tweaking the weights that determine certain datapoints to get more interesting results and correlation.