Finished MVP
The completed program (v0.5) is finished. This should meet all project requirements but it still feels pretty incomplete with what the foundation could be used for.

Tweaked tons of small things to provide more interesting graphs.
Deleted the sqlite file as it was completely unneeded.
Added a main loop to the program.